+++ ak_library_management_enhancement/models/author_author.py
@@ -19,25 +19,3 @@ class Author(models.Model):
         string = "status",
         default='new',
     )
-
-    def _get_book_domain(self):
-        self.ensure_one()
-        self.total_book = self.env['library.book'].search_count([('author_id', '=', self.id)])
-        return [('author_id', '=', self.id)]
-
-    def book_info(self):
-        self.ensure_one()
-        return {
-            'type': 'ir.actions.act_window',
-            'res_model': 'library.book',
-            'view_mode': 'list,form',
-            'domain': self._get_book_domain(),
-            'context': {
-                'default_author_id': self.id,
-            },
-            'target': 'current',
-        }
-
-    def total_book_count(self):
-        self.total_book = self.env['library.book'].search_count([('author_id', '=', self.id)])
-