+++ ak_library_management_orm/models/product_category.py
@@ -1,4 +1,5 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
+# -*- coding: utf-8 -*-
+
 from odoo import api, fields, models
 
 
@@ -9,17 +10,26 @@ class ProductCategory(models.Model):
         comodel_name='book.category',
         string = 'book product',
     )
+    total_book_categories = fields.Integer(
+        string="Categories",
+        compute="_compute_book_categories_count"
+    )
 
-    def action_open_book_view(self):
+    def action_open_book_category_view(self):
         """
         Redirect to the of the related book.category form view.
         """
-        print('\n\n',self.book_categ_id == None)
         return {
             'type': 'ir.actions.act_window',
             'res_model': 'book.category',
             'res_id': self.book_categ_id.id,
             'view_mode': 'form',
-            'context': {'active_id': self.id},
             'target': 'current',
         }
+
+    @api.depends('book_categ_id')
+    def _compute_book_categories_count(self):
+        """
+        Count the total book categories of the product category.
+        """
+        self.total_book_categories = len(self.book_categ_id)