+++ ak_library_management_orm/models/res_users.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class ResUsers(models.Model):
     _inherit = 'res.users'
     
-    is_manager = fields.Boolean(string='Manager',default=False)
+    is_manager = fields.Boolean(string='Manager')